problems/src/test/kotlin/data_structures/arrays/medium/MaxAreaOfIslandTest.kt
package data_structures.arrays.medium

import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Test

class MaxAreaOfIslandTest {
    @Test
    fun test1() {
        val grid = arrayOf(
            intArrayOf(0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0),
            intArrayOf(0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0),
            intArrayOf(0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0),
            intArrayOf(0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0),
            intArrayOf(0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0),
            intArrayOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0),
            intArrayOf(0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0),
            intArrayOf(0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0),
        )
        val expected = 6

        assertEquals(expected, MaxAreaOfIsland.maxAreaOfIsland(grid))
    }

    @Test
    fun test2() {
        val grid = arrayOf(
            intArrayOf(0, 0, 0, 0, 0, 0, 0, 0),
        )
        val expected = 0

        assertEquals(expected, MaxAreaOfIsland.maxAreaOfIsland(grid))
    }
}
